Add archive header with Empty Bin menu, fix checkbox and hover layout
Header:
- Create HeaderMainArchive with "..." more button containing Empty Bin
- Register mainArchive in header component mapping
- Move Empty Bin from title area to standard header position
- Only show "..." for Owner

Layout:
- Checkbox positioned absolutely outside the table grid (left: -36px)
- Remove 36px checkbox column from grid template
- Table has padding-left: 36px to make room for checkboxes
- Row hover uses ::before pseudo-element extending into checkbox area
  with border-radius: 8px

Column headers:
- Sortable column names have hover background (rounded, 6px radius)
- Padding adjusted for inline hover effect
